// frontend/src/sections/cold-chain-equipment/catalogue.generated.ts
export type AssetCategoriesListVariables = {
  filter?: {
    id?: {
    equalTo?: string | null;
    equalAny?: Array<string> | null;
    notEqualTo?: string | null;
    equalAnyOrNull?: Array<string> | null;
    notEqualAll?: Array<string> | null;
  } | null;
    name?: {
    equalTo?: string | null;
    like?: string | null;
  } | null;
    classId?: {
    equalTo?: string | null;
    equalAny?: Array<string> | null;
    notEqualTo?: string | null;
    equalAnyOrNull?: Array<string> | null;
    notEqualAll?: Array<string> | null;
  } | null;
  } | null;
};

export type AssetCategoriesListResult = {
  assetCategories: ({
  __typename: "AssetCategoryConnector";
} & {
  totalCount: number;
  nodes: Array<{
  id: string;
  name: string;
  classId: string;
}>;
});
};

export const AssetCategoriesList = {
  query: "query assetCategoriesList($filter: AssetCategoryFilterInput) {\n  assetCategories(filter: $filter) {\n    ... on AssetCategoryConnector {\n      __typename\n      totalCount\n      nodes {\n        id\n        name\n        classId\n      }\n    }\n  }\n}",
} as TypedDocument<AssetCategoriesListResult, AssetCategoriesListVariables>;

export type AssetTypesListVariables = {
  filter?: {
    id?: {
    equalTo?: string | null;
    equalAny?: Array<string> | null;
    notEqualTo?: string | null;
    equalAnyOrNull?: Array<string> | null;
    notEqualAll?: Array<string> | null;
  } | null;
    name?: {
    equalTo?: string | null;
    like?: string | null;
  } | null;
    categoryId?: {
    equalTo?: string | null;
    equalAny?: Array<string> | null;
    notEqualTo?: string | null;
    equalAnyOrNull?: Array<string> | null;
    notEqualAll?: Array<string> | null;
  } | null;
  } | null;
};

export type AssetTypesListResult = {
  assetTypes: ({
  __typename: "AssetTypeConnector";
} & {
  totalCount: number;
  nodes: Array<{
  id: string;
  name: string;
  categoryId: string;
}>;
});
};

export const AssetTypesList = {
  query: "query assetTypesList($filter: AssetTypeFilterInput) {\n  assetTypes(filter: $filter) {\n    ... on AssetTypeConnector {\n      __typename\n      totalCount\n      nodes {\n        id\n        name\n        categoryId\n      }\n    }\n  }\n}",
} as TypedDocument<AssetTypesListResult, AssetTypesListVariables>;